Initial commit with basic ENTTEC widget support.
Basically it's just an application skeleton with nothing more than the
implementation of the DMX widget ABI of the ENTTEC DMX USB PRO Mk2 and
is thus the first commit for the successor of GrandPA, a very quick and
dirty implementaion of a LED stick controller:

https://github.com/aszlig/GrandPA

I've already started to work on a reimplementation of GrandPA in Haskell
a while ago, but after some experimentation I decided to only do the UI
in Haskell and the backend in Erlang, because it's properties fit much
better when working with lighting fixtures (fault tolerance, distributed
architecture, code hot-reloading).

Right now, it's only a gen_fsm sending some widget test data. DMX data
is received on port 1 and sent on port 2. The reason for this is that
input/output ports are electrically paired and thus can act only in one
direction.

One might not think this is a problem and think that we can just send
DMX data and switch the port direction back to receiving mode again.

Unfortunately, if we're sending DMX data the port direction is set to
output only, which means that every incoming DMX packet is also passed
through to the output port as well, thus corrupting the output stream.

In order to get access to the second port on the DMX USB PRO Mk2, you
have to request an API key (shouldn't it be ABI?) from ENTTEC. I got a
key and also permission to publish it along with the source code.

Speaking of the source code itself:

This is published under the terms of the GNU Affero General Public
License, because I want to ensure that all modifications of this code
stay free, even when used on a remote server.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>