Refactor widgets to use a router module for DMX
We now have every DMX input and output in one location, the router.

Every widget now has a rayactor_widget behaviour set, so it's unified to
properly interface with the router.

All the startup handling and pid passing is done within the
rayactor_widget behaviour as well, so that only a minimum amount of
information and boilerplate is required on the widget side.

The only information a widget now has to care about is its ports and the
startup function, everything else is passed through the supervisor
children specification.

Right now the whole implementation is a bit of a mixed bag of stuff that
still needs to be cleaned up a bit, but so far the router works well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>