Add a small library for a dmx() data type
Currently, this is implemented on Erlang binaries and only has one
helper function set() and nothing more so far (more coming soon of
course).

The goal is to only use this library for handling DMX packets,
especially if it comes to only changing a subset of a packet.

We have to see whether using Erlang binaries is sufficient enough in
terms of performance and memory usage and probably implement it as a
NIF at some point in the future.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>