router: Fix recv_dmx functionality
We need to use the "input_buffer" key, because it contains the latest
values of the specific input. The "inputs" key just contains process
information instead.

Of course, the implementation is currently a bit fragile and we need to
have _real_ recv and async_recv functionality instead of just asking for
a buffer.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>