router: Make more robust to widget failure
If a widget process goes down it is immediately started again by its
supervisor and calls register_widget again. This has the downside that
receivers may listen to the old ports.

So we now properly deregister the widget once the process terminates, so
that once the widget becomes alive again we get outputs/inputs for the
new pid.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>