Add initial implementation of a native UDP client
The goal is to have a native RayActor client, which shows a visual
representation of the stage and can also forward keyboard inputs and
other DMX inputs coming from the RayActor server.

So far the idea is to loop the DMX outputs from the router back to the
client so that an accurate visualisation of the stage can be shown.

Right now the client is doing nothing more than just connecting via UDP,
sending some arbitrary magic string and display some shapes based on
further packets we receive.

Things such as different ports and transmission of fixture data is
currently not implemented as well as proper keypress handling.

Another thing that's missing here is a better way to handle the main
event loop, because right now this is using 100% CPU because we don't
have a subscription to Time.fps and in theory we won't even need it,
because all we have is an event loop around network and keyboard events.
Unfortunately neither Helm nor SDL easily allow us to select/poll over
the raw file descriptors.

The implementation for netReceived is also a bit ugly, because I haven't
found a good to handle UDP with Haskell's network library without using
threads. Using socketToHandle would work in theory, but we'd lose access
to the Socket.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>