Introduce a new ftdi port driver
This one is not yet used by the ENTTEC widget but it's going to replace
the uart library, because the latter is relying on the kernel-provided
/dev/ttyUSB* serial device which not only is limited but also doesn't
provide certain functionalities. One functionality in particular is to
purge the widget queue.

The implementation now uses libftdi1 instead which in turn uses libusb1
directly instead of talking with /dev/ttyUSB*. Although quite messy and
undocumented at large, it works with a few tests so far. It also lacks
error handling in some places.

A bigger issue was that libftdi1 - albeit using the asynchronous API of
libusb1 - is wrapping everything into a ftdi_transfer_control structure
which in turn only has functions that block on waiting for data.

So in essence, libftdi1 is synchronous. We don't want that, so this
implementation contains our own wrapper (transfer.c) around the
internals of libftdi1 to facilitate asynchronous sending and receiving.

Unfortunately, we still have some synchronous bits in there. For example
the main reason for doing this implementation: purge.

If you're asking yourself why purge is so essential here: A lot of DMX
controllers push out DMX packets in a very high rate, which causes the
widget queue to overflow.

Also, we will get rid of two dependencies: dthread and uart :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>