ftdi_drv: Tag async messages with a reference
If we use send/2 or recv/2 we used to get a tuple like this:

  {ftdi, Data}

This doesn't work well if we want to associate such asynchronous data
with the initial request.

Now both functions return a reference which can later be matched via
receive and it also makes implementation of synchronous calls very
trivial (they just imply the receive expression after sending out the
request).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>