rayactor_enttec: Use new ftdi library
This now uses our now ftdi port driver for communicating with the
widget.

It still has rough edges, for example we don't have a unrecv
functionality in our port driver and probably won't need it.

However, what we need is a way to guarantee that we don't have queue
overflows during receive. For example if a lot of incoming DMX packets
are flowing in during wait_for_hw_version state, we could hit a queue
overflow.

In order to mitigate this we now have a purge command, but that's not
enough, because we still have a brief time window between the purge and
the receiving of the next package where new packets can fill up the
widget's queue.

So to address this issue, we probably need to have a recv_purge or
something called like that which makes receiving and purge an atomic
operation from the view of the emulator.

Another thing that got uglier now that we don't have unrecv anymore is
the matching of the body of widget packets, because we can't do
something like this in Erlang:

foo(Len, <<Data:Len/binary, ...>>) -> something.

What got better now is that we now can directly match on the serial
number for a specific device. So if we specify #{serial => <<"ABC">>} in
the application config, only the device with serial "ABC" gets matched
to a particular universe.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>