ftdi_drv: Build inside _build directory
We have different profiles for debugging and release, so for example if
we have compiled the default profile and want to switch to a debugging
build, the final .so won't be regenerated by the makefile in some cases.

There is already a _build directory created by rebar3, which includes
seperate subdirectories for each profile, which is what we're now using
for placing build objects and results.

Unfortunately rebar3 makes symlinks between the priv directory of an
application within the _build directory and the applications source code
outside of the _build directory. This introduces the same problem I
described in the first paragraph again.

The workaround for this is to remove the symlink and mkdir the
directory, which works because rebar3 doesn't create that symlink if the
path already exists.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>