Add a small default.nix for debugging
Currently this is only used for within nix-shell, because the Erlang
release handling implementation in nixpkgs is non-existent and even if
we override installPhase it's still useless because we can't use relx
anyway (the build then fails with a very helpful error:

escript: exception error: no case clause matching ["relx","exe","3.18.0"]

So at first we really need to fix the broken Erlang support that we
currently have in nixpkgs before we can make releases in a pain-free
way.

The integration of valgrind overrides Erlang by compiling the special
beam.valgrind.* binaries, which make using valgrind a bit more painless
because by default running an emulator will flood you with false
positives.

There is also a fullDebug argument which compiles the whole dependency
graph up to rayactor with debugging info so we can *really* inspect
what's going on (event though I didn't need it so far, but it might come
in handy some day).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>